Do not allow a user to join a lab session twice
This adds a unique index on the lab session memberships for the user id
and the lab session id. This will not allow a user to be in the lab
session membership twice. If join is requested again once a user is
already in the lab session, it will just find the one that already
exists and it will return the same object.