Standardize errors that are produced
Generally there will be two different types of errors that can occur. 1
is that a user cannot find (or access) a record of a certain type. The
other is that there are improperly formatted or validated inputs. This
adds error handling for both of these cases in the general case so that
we can reuse them across all controllers and they are standardized.