Finish CRUD actions for lab sessions
This allows users to index, show, update, and destroy lab sessions. The
users must be a part of the lab session to perform any of these
actions. Additionally, for deleting a lab session, the user must either
be the only one on the session or they must be a professor.